Moved RHGPIO.py to 'util' directory
